+++ src/renderer/components/common/index.ts
@@ -1,0 +1,8 @@
+/**
+ * Common Components Index
+ *
+ * Re-exports all shared/common components.
+ */
+
+export { LogViewer } from './LogViewer'
+export { ThemeToggle } from './ThemeToggle'